CHANGED:  - Usando o padrão DAO DELETED  - Removido PersistenceManager (tava feio)
